requirements.txt.txt
pillow==9.2.0
pycryptodomex==3.18.0
tkintertable
